refactor: move browser-specific storage code to its own file
