+++ src/content/utils.ts
@@ -2,8 +2,6 @@ import { ISectionData, Term } from "./App/App.types"
 import { findCourseInfo } from "../backends/scheduler/nameSearchApi"
 import { fetchWorkdayData } from "../backends/workday/idSearchApi"
 import { handleCourseLoading } from "."
-import { handleProgressUpdate as modalProgressUpdater } from "../content/Settings/ExportImport/ExportImport"
-import { handleProgressUpdate as mainProgressUpdater } from "../backends/scheduler/nameSearchHelpers"
 
 // bypassDetailsCheck is for reskin extension compat
 async function extractSection(element: Element, bypassDetailsCheck?: boolean) {
@@ -102,14 +100,4 @@ const versionOneFiveZeroUpdateNotification = () => {
     .catch((error) => console.error("Error retrieving flag:", error))
 }
 
-const sendProgressUpdateToAll = (update: number) => {
-  modalProgressUpdater(update)
-  mainProgressUpdater(update)
-}
-
-export {
-  versionOneFiveZeroUpdateNotification,
-  filterSections,
-  extractSection,
-  sendProgressUpdateToAll,
-}
+export { versionOneFiveZeroUpdateNotification, filterSections, extractSection }
